added doc to explain how to add one more provider
